requirements.txt
fastapi
uvicorn
transformers
faiss-cpu
scikit-learn
llama-cpp-python==0.2.24
torch
numpy==1.24.4
python-multipart
gradio
pdfplumber
python-docx
openpyxl
pandas
piper-tts
# The above code is a requirements file for a Python project that uses FastAPI, Uvicorn, and Gradio for building a web application.
# It also includes libraries for machine learning, document processing, and data manipulation.
# The specific versions of some libraries are specified to ensure compatibility and stability.
# The code is structured to be modular and can be extended with additional features or improvements as needed.
# The requirements file is essential for setting up the project environment and ensuring that all necessary dependencies are installed.
# The project is designed to be user-friendly and provides a web interface for interacting with the functionalities.
# The app is intended to be run locally and can be accessed via a web browser.
# The code is designed to be efficient and optimized for performance, making use of modern libraries and frameworks.
# The project is suitable for developers and data scientists who want to build applications that involve document processing, machine learning, and data analysis.
# The requirements file is a crucial part of the project setup and should be maintained to reflect any changes in dependencies or versions.
# The project is expected to evolve over time, and the requirements file should be updated accordingly to ensure compatibility with new features or improvements.
# The code is designed to be clean and well-organized, following best practices for Python development.
# The project is open to contributions and collaboration, and the requirements file serves as a starting point for setting up the development environment.
# The project is intended to be user-friendly and provides a web interface for interacting with the functionalities.
# The app is designed to be run locally and can be accessed via a web browser.